[bug/59655] Use lang parameter instead of sid for language selection.
Current the SID is passed with each request to style.php this ruins
the usefulness of shared network level caches and also pollutes the
users client cache with a new stylesheet for each session they have
on the board. I've removed the internal dependency on the SID and
instead use the lang parameter to select the correct language to use.
Problem now is with a the same URI for each style/language combination
the automatic purging of the cached style on session ID change is
removed which means we need add an extra parameter to the URI reflecting
the modification time of the stylesheet being delivered.